defconfig: disable watchdog
watchdog timer support is expensive and is mainly just to restart a device
if there's a lockup when writing to a file.  I don't believe this is very applicable
to Android because in the extremely unlikely event that the device locks up.. the
user can just restart it manually.  It seems like a feature more suited for automated
systems and network systems that need to come back online as fast as possible.